refactor(main_logic): split omni realtime client into package
