+++ main_logic/omni_realtime_client/_responses.py
@@ -1,0 +1,680 @@
+# -- coding: utf-8 --
+# Copyright 2025-2026 Project N.E.K.O. Team
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from ._shared import (
+    Any,
+    Callable,
+    Dict,
+    Optional,
+    asyncio,
+    base64,
+    logger,
+    np,
+    soxr,
+    time,
+    uuid,
+)
+
+from ._proactive_audio import (
+    _load_proactive_audio,
+)
+
+
+class _ResponseMixin:
+    async def prime_context(self, text: str, skipped: bool = False) -> None:
+        """Inject context during hot-swap.
+
+        Behaviour depends on the skipped parameter and the provider:
+
+        - ``skipped=True`` (or Qwen): appended to the system instructions
+          via ``session.update``, without triggering a model response.
+        - ``skipped=False`` (GPT/GLM/Step): injects a one-shot user message
+          via ``create_response`` and triggers a model response (used for
+          proactively reporting task results). Note: this path does not
+          write to session instructions; the text is transient — do not
+          change it to persist into instructions.
+        - Gemini: injected via ``send_client_content`` regardless of
+          skipped (SDK limitation, no session.update mechanism). When
+          skipped=True the response is silently discarded via
+          ``_skip_until_next_response``.
+
+        Args:
+            text: Context to inject (incremental cache + summary/ready).
+            skipped: If True, only update instructions without triggering
+                     a response. If False, also trigger model response.
+        """
+        if not text or not text.strip():
+            logger.info("prime_context: skipping empty content")
+            return
+
+        if self._is_gemini:
+            # Gemini Live API 没有 session.update 机制，只能通过
+            # send_client_content 注入上下文（会创建 user turn）。
+            # on_response_done 由 _handle_messages_gemini 自然触发。
+            await self._create_response_gemini_with_skip_guard(
+                text,
+                skipped=skipped,
+                raise_on_error=True,
+            )
+            return
+
+        if not skipped and "qwen" not in self._model_lower:
+            # skipped=False：需要模型主动响应（任务结果汇报）
+            # 通过 create_response 注入 user 消息 + 触发响应
+            # Qwen 不支持 conversation.item.create，走下方 update_session
+            await self.create_response(text)
+        else:
+            # skipped=True 或 Qwen：仅追加到 session instructions
+            lock = getattr(self, "_prime_context_lock", None)
+            if lock is None:
+                lock = asyncio.Lock()
+                self._prime_context_lock = lock
+            async with lock:
+                current_instructions = str(self.instructions or "")
+                next_instructions = (
+                    current_instructions + "\n" + text
+                    if current_instructions
+                    else text
+                )
+                await self.update_session({"instructions": next_instructions})
+                self.instructions = next_instructions
+            logger.info("prime_context: updated session instructions")
+
+    async def create_response(self, instructions: str, skipped: bool = False) -> None:
+        """Inject a persistent user message and trigger an LLM response.
+
+        Unlike ``prime_context`` (which appends to the system instructions),
+        this method creates a user-role conversation message and triggers a
+        model response. Suited to mid-conversation scenarios where an
+        immediate model reply is needed.
+
+        Note: requires that the session already contains a user message, or
+        that the API in use supports ``conversation.item.create``; otherwise
+        a 1007 error may be triggered.
+
+        Behaviour varies by provider:
+          - **OpenAI / GLM / Step**: ``conversation.item.create(role=user)``
+            + ``response.create``
+          - **Gemini**: ``send_client_content(role=user)``
+
+        See ``prime_context()`` (session-start priming) and
+        ``prompt_ephemeral()`` (fire-and-forget audio nudge) for the other
+        two injection channels.
+        """
+        # Gemini 使用 send_client_content 发送文本内容
+        if self._is_gemini:
+            if not instructions or not instructions.strip():
+                logger.info("Gemini: skipping empty content in create_response")
+                return
+            await self._create_response_gemini_with_skip_guard(
+                instructions,
+                skipped=skipped,
+                raise_on_error=True,
+            )
+            return
+
+        # 跳过空内容的发送，避免触发 API 错误
+        if not instructions or not instructions.strip():
+            logger.info("Skipping empty content in create_response")
+            return
+
+        if skipped:
+            self._skip_until_next_response = True
+
+        # 通过 conversation.item.create 添加用户消息，再触发响应
+        item_event = {
+            "type": "conversation.item.create",
+            "item": {
+                "type": "message",
+                "role": "user",
+                "content": [
+                    {
+                        "type": "input_text",
+                        "text": instructions
+                    }
+                ]
+            }
+        }
+        await self.send_event(item_event)
+
+        logger.info("Creating response with user message")
+        await self.send_event({"type": "response.create"})
+
+    def is_active_response(self) -> bool:
+        """Return True iff the realtime session is currently producing a response.
+
+        Tracks ``response.created`` → ``response.done`` (OpenAI / GLM / Step /
+        free / GPT) and Gemini's ``turn_complete`` lifecycle via the shared
+        ``_is_responding`` flag, so callers can gate "manual inject + request
+        response" against the realtime API's "one active response at a time"
+        constraint.
+        """
+        return bool(self._is_responding)
+
+    async def inject_text_and_request_response(
+        self,
+        text: str,
+        *,
+        on_rejected: Optional[Callable[[str], None]] = None,
+    ) -> None:
+        """Inject a user-role text item and explicitly trigger a response.
+
+        Used by the voice-mode proactive path (agent task callbacks /
+        plugin push_message ai_behavior="respond") to surface a rendered
+        instruction to the realtime model and have it speak the result
+        immediately — without waiting for the next user turn (which is what
+        the hot-swap pending_extra_replies channel does).
+
+        Caller is responsible for gating against active-response races
+        (see ``is_active_response``) — the realtime API only allows one
+        in-flight response at a time and will reject a second
+        ``response.create`` with ``response_already_active``.
+
+        Server-side rejection (e.g. VAD races in between the caller's gate
+        check and our ``response.create``) does not raise here because the
+        server delivers it asynchronously via an ``error`` event. Pass
+        ``on_rejected=cb(error_msg)`` to receive that rejection — the
+        message loop will invoke it when ``error.event_id`` matches the
+        client-side id we stamp on ``response.create``. The caller can use
+        it to put the optimistically-pruned cb back in the queue.
+
+        Provider dispatch (all realtime providers supported — symmetric with
+        ``create_response``):
+          - **OpenAI / GLM / Step / free / GPT / Qwen / Grok**:
+            ``conversation.item.create`` (role=user, input_text) +
+            ``response.create``. Uses user role rather than system to avoid
+            permanent drift of session instruction context — the rendered
+            body already self-identifies as a system notification via its
+            localized header wrapper. (Qwen included: the
+            Aliyun doc claiming function_call_output-only is stale for
+            qwen3.5-omni-flash-realtime; verified live.)
+          - **Gemini Live**: ``send_client_content(turn_complete=True)`` via
+            the shared ``_gemini_send_user_turn`` helper — Gemini's idiomatic
+            inject+trigger. No ``on_rejected`` async ack (Gemini has no
+            ``response.create`` error-event channel); failures raise
+            synchronously here so the caller's ``except`` branch re-queues.
+        """
+        if self._fatal_error_occurred:
+            raise RuntimeError("realtime session has fatal_error_occurred set")
+        if not text or not text.strip():
+            return
+
+        if self._is_gemini:
+            # Symmetric with create_response → _create_response_gemini.
+            # send_client_content(turn_complete=True) injects a user turn and
+            # triggers a response. Errors propagate (unlike the swallowing
+            # _create_response_gemini wrapper) so the caller keeps the cb.
+            if self._gemini_session is None:
+                raise RuntimeError("Gemini session not available for proactive inject")
+            await self._gemini_send_user_turn(text)
+            return
+        # NOTE on Qwen: the Aliyun realtime doc states conversation.item.create
+        # "currently only supports function_call_output items". That is stale
+        # for qwen3.5-omni-flash-realtime — empirically it accepts a
+        # ``role=user`` ``input_text`` message item and responds to it (no
+        # error event), identical to OpenAI / GLM / Step. Verified live against
+        # the dashscope realtime endpoint. So Qwen takes the same path below;
+        # do NOT re-add a Qwen exclusion without re-checking the live API.
+        if self.ws is None:
+            raise RuntimeError("realtime websocket is not connected")
+
+        # Role choice: ``user`` (not ``system``).
+        # OpenAI Realtime persists conversation items as part of session
+        # history. ``role="system"`` items are treated as high-priority
+        # instructions that influence every subsequent turn — accumulating
+        # several proactive callbacks under system role causes prompt
+        # drift (model starts repeating meta-behavior or interpreting
+        # stale callback text as standing orders for unrelated turns).
+        # ``role="user"`` keeps the inject in dialog-weight context, and
+        # ``_build_callback_instruction`` already wraps the body in a
+        # ``======[系统通知] ...======`` header that makes the model
+        # treat it as a one-shot system notification rather than user
+        # speech. Matches the existing ``create_response`` precedent.
+        # Stamp stable client event_ids on BOTH events so the server's
+        # ``error.event_id`` can be matched back to this specific request
+        # whichever event it rejects (the item itself, or the
+        # ``response.create`` — e.g. ``response_already_active`` from a VAD
+        # race). ``send_event()`` would otherwise overwrite a missing
+        # event_id with its own timestamp-based string — fine for routing but
+        # useless for rejection matching since the caller has no view of it.
+        # A single ``_reject_once`` wrapper fires ``on_rejected`` at most once
+        # even if both event_ids somehow error, and unregisters both handlers.
+        item_event_id: Optional[str] = None
+        create_event_id: Optional[str] = None
+        if on_rejected is not None:
+            item_event_id = f"event_inject_item_{uuid.uuid4().hex}"
+            create_event_id = f"event_inject_resp_{uuid.uuid4().hex}"
+
+            _fired = False
+
+            def _reject_once(error_msg: str) -> None:
+                nonlocal _fired
+                # Unregister both regardless so neither lingers.
+                self._inject_rejection_handlers.pop(item_event_id, None)
+                self._inject_rejection_handlers.pop(create_event_id, None)
+                if _fired:
+                    return
+                _fired = True
+                on_rejected(error_msg)
+
+            self._inject_rejection_handlers[item_event_id] = _reject_once
+            self._inject_rejection_handlers[create_event_id] = _reject_once
+            # The realtime API echoes our event_id on ``error`` but NOT on
+            # ``response.created`` — so a successful inject leaves the handlers
+            # registered with no natural cleanup signal. Primary cleanup is
+            # lifecycle-based: ``response.done`` sweeps the dict (see
+            # ``_sweep_inject_rejection_handlers`` — a rejection is always
+            # emitted before the blocking response completes, so any pending
+            # rejection has already fired by any response.done). This TTL is
+            # only a backstop for the pathological "no response.done ever"
+            # case (session hangs); 30s is generous vs the sub-second
+            # rejection latency, so a real ``response_already_active`` reject
+            # under transient backpressure is still caught (Codex P2).
+            self._fire_task(self._expire_inject_rejection_handler(item_event_id, 30.0))
+            self._fire_task(self._expire_inject_rejection_handler(create_event_id, 30.0))
+            # Open the no-id content-fallback window for THIS inject. Closed
+            # when its outcome is observed (rejection fired, or the next
+            # response lifecycle event / done sweep) — see
+            # _route_inject_rejection.
+            self._proactive_inject_awaiting_outcome = True
+
+        item_event: Dict[str, Any] = {
+            "type": "conversation.item.create",
+            "item": {
+                "type": "message",
+                "role": "user",
+                "content": [
+                    {
+                        "type": "input_text",
+                        "text": text,
+                    }
+                ],
+            },
+        }
+        if item_event_id is not None:
+            item_event["event_id"] = item_event_id
+
+        # send_event() silently returns when ws drops to None or fatal flag
+        # flips mid-flight (it does not raise). Without the post-send checks,
+        # a connection lost in the brief await window between the entry guard
+        # and the actual send would look like a successful inject — caller
+        # would prune the cb but nothing reached the model. Re-check after
+        # each send and raise so the caller's exception branch keeps the cb
+        # for retry. On any synchronous send failure, drop both rejection
+        # handlers so the caller's ``except`` path is the single source of
+        # truth and a late error event can't double-fire the re-queue.
+        try:
+            await self.send_event(item_event)
+            if self._fatal_error_occurred or self.ws is None:
+                raise RuntimeError(
+                    "realtime connection lost after proactive conversation.item.create"
+                )
+            create_event: Dict[str, Any] = {"type": "response.create"}
+            if create_event_id is not None:
+                create_event["event_id"] = create_event_id
+            await self.send_event(create_event)
+            if self._fatal_error_occurred or self.ws is None:
+                raise RuntimeError(
+                    "realtime connection lost after proactive response.create"
+                )
+        except Exception:
+            if item_event_id is not None:
+                self._inject_rejection_handlers.pop(item_event_id, None)
+            if create_event_id is not None:
+                self._inject_rejection_handlers.pop(create_event_id, None)
+            raise
+
+    async def _expire_inject_rejection_handler(self, event_id: str, ttl: float) -> None:
+        """TTL backstop cleanup for the inject rejection handler dict (see
+        ``inject_text_and_request_response``). Primary cleanup is the
+        lifecycle sweep in ``_sweep_inject_rejection_handlers``; this only
+        catches the pathological "no response.done ever" case."""
+        try:
+            await asyncio.sleep(ttl)
+        except asyncio.CancelledError:
+            return
+        self._inject_rejection_handlers.pop(event_id, None)
+
+    @staticmethod
+    def _looks_like_response_conflict(error_msg: str) -> bool:
+        """Heuristic: does this ``error`` message look like the server
+        rejecting a ``response.create`` because a response is already active?
+
+        That ``response_already_active`` class is the ONLY async rejection a
+        proactive inject can provoke (our inject is the only client-issued
+        ``response.create`` on the voice path). Matching its content lets us
+        route the rejection even when the provider's error doesn't echo our
+        client ``event_id``. Kept deliberately broad across phrasings /
+        providers but still scoped to response-conflict wording so unrelated
+        errors (auth / quota / 503 / idle-timeout) don't trip it."""
+        low = error_msg.lower()
+        if "response_already_active" in low:
+            return True
+        return "response" in low and any(
+            k in low for k in ("already", "active", "in progress", "in_progress", "exists", "ongoing")
+        )
+
+    def _route_inject_rejection(self, err_event_id, error_msg: str) -> None:
+        """Deliver a server rejection to the matching proactive-inject
+        ``on_rejected`` handler so the caller re-enqueues the cb.
+
+        Two correlation paths:
+          1. **By id (precise)** — OpenAI Realtime (and any provider that
+             echoes the offending client ``event_id``): pop and fire the exact
+             handler.
+          2. **By content (fallback)** — ONLY when the provider omits a
+             client-correlation id entirely (``err_event_id`` falsy): if the
+             error looks like a response-conflict
+             (``_looks_like_response_conflict``) and handlers are pending, fire
+             them all. Injects are serialized by
+             ``_voice_proactive_inject_lock`` so there's effectively one logical
+             pending inject; ``_reject_once`` + the caller's delivery-id dedup
+             make a spurious fire cost at most a bounded duplicate re-add, which
+             is strictly better than a silent drop (Codex P1).
+
+        Critically, the content fallback is gated on ``err_event_id`` being
+        absent. If the error DOES carry a client event_id that simply isn't
+        ours, the rejection belongs to a different ``response.create`` (e.g.
+        ``create_response`` hot-swap priming / tool-result continuation /
+        ``signal_user_activity_end`` — all of which get a timestamp event_id
+        from ``send_event``'s setdefault), NOT our inject. Firing our handlers
+        on those would re-enqueue callbacks the model actually accepted →
+        duplicate announcements. So a present-but-non-matching id means "not
+        ours; do nothing"."""
+        if not self._inject_rejection_handlers:
+            return
+
+        def _fire(handler) -> None:
+            try:
+                handler(error_msg)
+            except Exception as cb_exc:
+                logger.warning("proactive inject rejection handler raised: %s", cb_exc)
+
+        if err_event_id:
+            # Id present: fire ONLY on an exact match. A non-matching id
+            # belongs to some other request's rejection — not ours.
+            handler = self._inject_rejection_handlers.pop(err_event_id, None)
+            if handler is not None:
+                self._proactive_inject_awaiting_outcome = False
+                _fire(handler)
+            return
+
+        # No client-correlation id at all — fall back to content matching,
+        # but ONLY while a proactive inject is genuinely awaiting its outcome
+        # (one-shot window). This excludes a no-id response-conflict raised by
+        # a DIFFERENT response.create sender (create_response / tool-result /
+        # signal_user_activity_end) from hitting a lingering, already-succeeded
+        # proactive handler.
+        if (
+            self._proactive_inject_awaiting_outcome
+            and self._looks_like_response_conflict(error_msg)
+        ):
+            self._proactive_inject_awaiting_outcome = False
+            for handler in list(self._inject_rejection_handlers.values()):
+                _fire(handler)
+            self._inject_rejection_handlers.clear()
+
+    def _sweep_inject_rejection_handlers(self) -> None:
+        """Drop all pending inject rejection handlers on a ``response.done``
+        lifecycle boundary.
+
+        (Only the WS-realtime ``response.done`` path calls this — the Gemini
+        branch of ``inject_text_and_request_response`` returns early via
+        ``_gemini_send_user_turn`` and never registers rejection handlers, so
+        Gemini's turn-complete has nothing to sweep.)
+
+        Safe because a server rejection of our ``response.create`` /
+        ``conversation.item.create`` is emitted the instant the server
+        receives a request it can't honor — and the only reason it can't
+        honor a ``response.create`` is that another response is already
+        active. That blocking response's ``response.done`` is therefore
+        strictly LATER than the rejection. So by the time ANY response.done
+        arrives, every pending rejection for a prior send has already fired
+        (and its handler self-removed via ``_reject_once``). Whatever remains
+        in the dict belongs to an inject that SUCCEEDED (no rejection coming)
+        — exactly the leak the fixed TTL was meant to clean, now reaped
+        promptly and lifecycle-tied instead of on a wall clock."""
+        # The inject's outcome has been observed (a response completed), so
+        # close the no-id content-fallback window too.
+        self._proactive_inject_awaiting_outcome = False
+        if self._inject_rejection_handlers:
+            self._inject_rejection_handlers.clear()
+
+    async def prompt_ephemeral(
+        self,
+        instruction: str = "",
+        *,
+        language: str = "zh",
+    ) -> bool:
+        """Send a fire-and-forget audio nudge to trigger proactive AI speech.
+
+        Injects a short WAV clip via ``input_audio_buffer.append`` so the
+        realtime model "hears" a conversational nudge and responds.  Bypasses
+        ``stream_audio()`` (no RNNoise / AGC) since the audio is clean.
+
+        Unlike ``prime_context`` (session-start system-prompt injection) and
+        ``create_response`` (persistent mid-conversation message), this
+        channel is truly ephemeral — the audio prompt is consumed by the
+        model but never stored in conversation history.
+
+        Chunk pacing mirrors hot-swap flush: 1600 bytes/chunk, 0.025 s sleep,
+        40 chunks/s → 2× real-time delivery.
+
+        Returns True if the audio was fully sent, False if skipped or aborted.
+        """
+        # ── Guard checks ──────────────────────────────────────────────
+        if self._fatal_error_occurred or self.ws is None:
+            return False
+        if self._is_responding:
+            logger.debug("prompt_ephemeral: skipped — already responding")
+            return False
+        _now = time.time()
+        # ── AI-speech guard（对称于 _user_recent_activity_time）─────────
+        # _is_responding 已被 response.done / turn_complete flip False，但 AI 侧
+        # content 流可能还在滴水：
+        #   1. Gemini turn_complete 早于最后几帧音频送达
+        #   2. Gemini 长回复 sub-turn 间的 False 瞬间
+        #   3. response.created 到首 content chunk 的空窗（_is_responding 已 True
+        #      覆盖这一条，但加这层冗余保险无害）
+        # 3s 窗口覆盖上述抢跑 gap，避免 fudge 踩着 AI 尾巴打断自己。
+        if _now - self._ai_recent_activity_time < self._ai_recent_activity_window:
+            logger.debug("prompt_ephemeral: skipped — AI recently active (%.2fs ago)",
+                         _now - self._ai_recent_activity_time)
+            return False
+        # ── User-speech guards ───────────────────────────────────────
+        # B: 先用独立的 _user_recent_activity_time 判定近期是否有语音帧；
+        # 此信号不依赖 sustain，覆盖用户说话首 500ms 与句间停顿缝隙。
+        # 适用所有 VAD 源（RNNoise / server-VAD / RMS），所以不再门控在
+        # _rnnoise_vad_active 下 —— RMS 阈值 500 已较保守，误触可接受，
+        # 相比"fudge 切断用户说话"的体验损失值得。
+        if _now - self._user_recent_activity_time < self._user_recent_activity_window:
+            logger.debug("prompt_ephemeral: skipped — user recently active (%.2fs ago)",
+                         _now - self._user_recent_activity_time)
+            return False
+        # A: 现有 _client_vad_active + grace 检查（sustained VAD 信号兜底）。
+        # Grace 已从 2s 扩到 6s，覆盖自然停顿。
+        # 门控条件：存在可靠 VAD 信号源。
+        #   - server-VAD 后端（Qwen/OpenAI）：server 的 speech_started/stopped 可靠，
+        #     不依赖 RNNoise。特别覆盖 16kHz 移动端长句 >8s 的场景（_user_recent_activity
+        #     在 speech_started 打点后 8s 过期，而用户还在说，需要 _client_vad_active 兜底）。
+        #   - RNNoise 客户端 VAD（48kHz 桌面 + Gemini/lanlan.app+free）
+        # RMS-only 路径（16kHz 无 server-VAD）信号太噪，不信任，依赖 _user_recent_activity。
+        if self._has_server_vad or self._rnnoise_vad_active:
+            if self._client_vad_active:
+                logger.debug("prompt_ephemeral: skipped — user speaking (VAD active)")
+                return False
+            if _now - self._client_vad_last_speech_time < self._client_vad_grace_period:
+                logger.debug("prompt_ephemeral: skipped — VAD grace period")
+                return False
+
+        # ── Choose audio file ─────────────────────────────────────────
+        # Vision context exists if an image was analyzed this turn (via
+        # VISION_MODEL text description OR native image input) or we have
+        # an unconsumed frame from stream_image().
+        has_vision = self._image_recognized_this_turn or (
+            self._latest_image_b64 is not None and not self._proactive_image_consumed
+        )
+        # Only backends with native image support can receive raw screenshots;
+        # step / lanlan.tech+free consume vision context as text only.
+        can_inject_image = has_vision and self._supports_native_image
+
+        # Snapshot the current image so concurrent stream_image() calls don't
+        # cause us to mark a newer frame as consumed.
+        snapshot_image_b64 = self._latest_image_b64 if has_vision else None
+
+        prompt_type = "vision" if has_vision else "general"
+        lang = (language or "zh")[:2]
+        filename = f"prompt_{prompt_type}_{lang}.wav"
+
+        try:
+            pcm_data = _load_proactive_audio(filename)
+        except FileNotFoundError:
+            try:
+                pcm_data = _load_proactive_audio(f"prompt_{prompt_type}_zh.wav")
+            except FileNotFoundError:
+                logger.warning("prompt_ephemeral: no audio file found for %s", filename)
+                return False
+
+        # Proactive WAVs are stored at 16kHz; for OpenAI's 24kHz-only uplink,
+        # upsample the whole clip once (stateless — it's a complete signal, so
+        # no chunk-boundary artifacts and no need to touch the mic-stream
+        # resampler). No-op for every 16kHz-native provider.
+        if self._uplink_sample_rate != 16000:
+            _clip = np.frombuffer(pcm_data, dtype=np.int16).astype(np.float32) / 32768.0
+            _clip = soxr.resample(_clip, 16000, self._uplink_sample_rate, quality='HQ')
+            pcm_data = (_clip * 32768.0).clip(-32768, 32767).astype(np.int16).tobytes()
+
+        # ── Non-native vision: inject text description before audio ───
+        # step / lanlan.tech+free can't receive raw images; send the
+        # VISION_MODEL text analysis so the model has visual context.
+        if has_vision and not can_inject_image and self._image_recognized_this_turn and self._image_description:
+            await self.send_event({
+                "type": "conversation.item.create",
+                "item": {
+                    "type": "message",
+                    "role": "user",
+                    "content": [{"type": "input_text", "text": self._image_description}],
+                },
+            })
+            logger.info("prompt_ephemeral: injected vision text description for non-native backend")
+
+        # ── Suppress mic input during injection ────────────────────────
+        self._proactive_injecting = True
+
+        # ── Send audio chunks (same pacing as hot-swap flush) ─────────
+        # 10 ms @16-bit mono = (rate/100)*2 bytes, ×5 multiplier → 50 ms/chunk.
+        # Rate-derived so pacing stays 50 ms/chunk after the 24kHz upsample.
+        chunk_size = (self._uplink_sample_rate // 100) * 2 * 5  # 50 ms of audio
+        sleep_interval = 0.025  # 25 ms → 40 chunks/s, 2× real-time
+
+        logger.info(
+            "prompt_ephemeral: injecting %s (%d bytes, %s)",
+            filename, len(pcm_data), "vision" if has_vision else "general",
+        )
+
+        total_chunks = (len(pcm_data) + chunk_size - 1) // chunk_size
+        mid_chunk = total_chunks // 2  # Insert image at the midpoint
+        image_injected = False
+
+        try:
+            _inject_start = time.time()
+            for chunk_idx, i in enumerate(range(0, len(pcm_data), chunk_size)):
+                # Abort conditions:
+                #   - AI started responding (self-interrupt protection)
+                #   - _client_vad_active sustained-speech fired (RNNoise only)
+                #   - B: any VAD source detected a new speech frame SINCE injection started
+                #     —— 注入过程中用户突然开口也能丢弃残余 chunk，不至于把用户
+                #     语音与 fudge 音频混在一起喂给模型
+                if self._is_responding or (self._rnnoise_vad_active and self._client_vad_active):
+                    logger.info("prompt_ephemeral: aborted — user spoke or response started")
+                    await self.clear_audio_buffer()
+                    return False
+                if self._user_recent_activity_time > _inject_start:
+                    logger.info("prompt_ephemeral: aborted — user started speaking during injection")
+                    await self.clear_audio_buffer()
+                    return False
+                # Gemini 首 content chunk 到达前 _is_responding 仍是 False（上面那条
+                # 拦不住），但 _ai_recent_activity_time 会在首 chunk 抵达瞬间更新到
+                # > _inject_start，此时 abort 避免和刚起的 AI 响应抢麦。
+                if self._ai_recent_activity_time > _inject_start:
+                    logger.info("prompt_ephemeral: aborted — AI started responding during injection")
+                    await self.clear_audio_buffer()
+                    return False
+
+                chunk = pcm_data[i : i + chunk_size]
+                if self._is_gemini:
+                    if self._gemini_session:
+                        await self._gemini_session.send_realtime_input(
+                            audio={"data": chunk, "mime_type": "audio/pcm"}
+                        )
+                else:
+                    audio_b64 = base64.b64encode(chunk).decode()
+                    await self.send_event({
+                        "type": "input_audio_buffer.append",
+                        "audio": audio_b64,
+                    })
+
+                # Inject cached screenshot at midpoint (only for native-image backends)
+                if can_inject_image and not image_injected and chunk_idx >= mid_chunk and snapshot_image_b64:
+                    if self._is_gemini:
+                        if self._gemini_session:
+                            image_bytes = base64.b64decode(snapshot_image_b64)
+                            await self._gemini_session.send_realtime_input(
+                                media={"data": image_bytes, "mime_type": "image/jpeg"}
+                            )
+                    elif "gpt" in self._model_lower:
+                        await self.send_event({
+                            "type": "conversation.item.create",
+                            "item": {
+                                "type": "message",
+                                "role": "user",
+                                "content": [{
+                                    "type": "input_image",
+                                    "image_url": "data:image/jpeg;base64," + snapshot_image_b64,
+                                }],
+                            },
+                        })
+                    elif "qwen" in self._model_lower or self._is_free_proxy:
+                        await self.send_event({
+                            "type": "input_image_buffer.append",
+                            "image": snapshot_image_b64,
+                        })
+                    elif "glm" in self._model_lower:
+                        await self.send_event({
+                            "type": "input_audio_buffer.append_video_frame",
+                            "video_frame": snapshot_image_b64,
+                        })
+                    image_injected = True
+                    logger.info("prompt_ephemeral: injected screenshot at chunk %d/%d", chunk_idx, total_chunks)
+
+                await asyncio.sleep(sleep_interval)
+
+            # Mark vision context consumed only if the shared image hasn't been
+            # replaced by a newer frame from stream_image() during our async loop.
+            if has_vision and self._latest_image_b64 == snapshot_image_b64:
+                self._proactive_image_consumed = True
+            logger.info("prompt_ephemeral: audio injection complete (%s%s), waiting for VAD → response",
+                         "vision" if has_vision else "general",
+                         "+image" if image_injected else "")
+            return True
+        finally:
+            self._proactive_injecting = False
+
+    async def cancel_response(self) -> None:
+        """Cancel the current response."""
+        event = {
+            "type": "response.cancel"
+        }
+        await self.send_event(event)
